+++ Editor/Apis/BucketsApiRequests.cs
@@ -1065,7 +1065,7 @@ namespace Unity.Services.Ccd.Management.Buckets
         /// <param name="path">Path</param>
         /// <param name="includeStates">Include change states, one or more can be specified. The default is all states.</param>
         /// <param name="label">Label</param>
-        /// <param name="contentType">conent type</param>
+        /// <param name="contentType">content type</param>
         /// <param name="complete">is content upload completed or not</param>
         /// <param name="sortBy">Sort By</param>
         /// <param name="sortOrder">Sort Order</param>
@@ -1254,7 +1254,7 @@ namespace Unity.Services.Ccd.Management.Buckets
         /// <param name="path">Path</param>
         /// <param name="includeStates">Include change states, one or more can be specified. The default is all states.</param>
         /// <param name="label">Label</param>
-        /// <param name="contentType">conent type</param>
+        /// <param name="contentType">content type</param>
         /// <param name="complete">is content upload completed or not</param>
         /// <param name="sortBy">Sort By</param>
         /// <param name="sortOrder">Sort Order</param>